Move DataMixingSiPixelMCDigiWorker to PreMixingSiPixelWorker
